# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Financial transaction processing and portfolio balance management microservice for GlobeCo benchmarking suite.\n\nThis service processes financial transactions and maintains portfolio account balances with:\n- Transaction creation and processing with comprehensive validation\n- Balance calculation and portfolio summary generation\n- Batch transaction processing for file imports\n- Real-time balance updates with optimistic locking\n- Integration with portfolio and security services",
        "title": "GlobeCo Portfolio Accounting Service API",
        "contact": {
            "name": "GlobeCo Support",
            "url": "https://github.com/kasbench/globeco-portfolio-accounting-service",
            "email": "noah@kasbench.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8087",
    "basePath": "/api/v1",
    "paths": {
        "/balance/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a specific balance record using its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balances"
                ],
                "summary": "Get balance by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Balance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved balance",
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid balance ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Balance not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/balances": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve portfolio balances with optional filtering by portfolio, security, and quantity ranges. Supports pagination and sorting.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balances"
                ],
                "summary": "Get balances with filtering",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by portfolio ID (24 characters)",
                        "name": "portfolio_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by security ID (24 characters). Use 'null' for cash balances",
                        "name": "security_id",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Pagination offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of records to return (default: 50, max: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort fields (comma-separated): portfolio_id,security_id",
                        "name": "sortby",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved balances",
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/portfolios/{portfolioId}/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a comprehensive summary of a portfolio including cash balance and all security positions with market values and statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balances"
                ],
                "summary": "Get portfolio summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio ID (24 characters)",
                        "name": "portfolioId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved portfolio summary",
                        "schema": {
                            "$ref": "#/definitions/dto.PortfolioSummaryDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid portfolio ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Portfolio not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.BalanceDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "portfolioId": {
                    "type": "string"
                },
                "quantityLong": {
                    "type": "number"
                },
                "quantityShort": {
                    "type": "number"
                },
                "securityId": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.BalanceListResponse": {
            "type": "object",
            "properties": {
                "balances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BalanceDTO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                }
            }
        },
...
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "dto.PortfolioSummaryDTO": {
            "type": "object",
            "properties": {
                "cashBalance": {
                    "type": "number"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "portfolioId": {
                    "type": "string"
                },
                "securities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SecurityPositionDTO"
                    }
                },
                "securityCount": {
                    "type": "integer"
                }
            }
        },
        "dto.SecurityPositionDTO": {
            "type": "object",
            "properties": {
                "lastUpdated": {
                    "type": "string"
                },
                "netQuantity": {
                    "type": "number"
                },
                "quantityLong": {
                    "type": "number"
                },
                "quantityShort": {
                    "type": "number"
                },
                "securityId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for service authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}